requirements.txt
fastapi==0.123.10
uvicorn==0.38.0
requests==2.32.5
beautifulsoup4==4.14.3
jsbeautifier==1.15.4
tree-sitter==0.25.2
tree-sitter-javascript==0.25.0
playwright==1.56.0
openai==2.9.0
tiktoken==0.12.0
google-auth==2.43.0
gspread==6.2.1
pdfkit==1.0.0
